requirements.txt
streamlit
numpy
pillow
scipy
opencv-python
scikit-image
tensorflow==2.3.0
imgaug
matplotlib
labelme2coco
pixellib==0.6.1